+++ src/components/index.ts
@@ -30,7 +30,7 @@ export {
   Search,
 } from "./Input";
 export { default as LayoutGridCols } from "./LayoutGridCols";
-export { default as LinkButton } from "./Link";
+export { default as LinkButton } from "./LinkButton";
 export { default as Navigation } from "./Navigation";
 export { default as Section } from "./Section";
 export { default as Table } from "./Table";
